Fixing chat redundant code
