+++ src/App.jsx
@@ -3,11 +3,11 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import SignIn from "./components/SignIn";
 import SignUp from "./components/SignUp";
-import StuHome from "./components/StuHome";
+import StudentHome from "./components/StudentHome";
 import Home from "./components/Home";
 import Projects from "./components/Projects";
 import Tasks from "./components/Tasks";
-import Chat from "./components/Chat";
+import AdminChat from "./components/AdminChat";
 import StudentTask from "./components/StudentTask";
 import StudentsChat from "./components/StudentsChat";
 import { DarkModeProvider } from "./Context/DarkModeContext";
@@ -23,8 +23,8 @@ function App() {
           <Route path="/home" element={<Home />} />
           <Route path="/projects" element={<Projects />} />
           <Route path="/tasks" element={<Tasks />} />
-          <Route path="/chat" element={<Chat />} />
-          <Route path="/stu-home" element={<StuHome />} />
+          <Route path="/chat" element={<AdminChat />} />
+          <Route path="/stu-home" element={<StudentHome />} />
           <Route path="/student-task" element={<StudentTask />} />
           <Route path="/student-chat" element={<StudentsChat />} />
         </Routes>
