+++ src/components/shared/index.js
@@ -1,0 +1,2 @@
+// Export all shared components for easier imports
+export { default as ChatInterface } from './ChatInterface';
